requirements.txt
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
et-xmlfile==1.1.0
idna==3.4
lxml==4.9.2
numpy==1.24.2
openpyxl==3.1.0
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.14
